Moved logic inside stop command and added flag
