+++ cli/index.ts
@@ -21,7 +21,6 @@ import { registerCommand as registerSiteSetWpVersionCommand } from 'cli/commands
 import { registerCommand as registerSiteStartCommand } from 'cli/commands/site/start';
 import { registerCommand as registerSiteStatusCommand } from 'cli/commands/site/status';
 import { registerCommand as registerSiteStopCommand } from 'cli/commands/site/stop';
-import { registerCommand as registerSiteStopAllCommand } from 'cli/commands/site/stop-all';
 import { commandHandler as wpCliCommandHandler } from 'cli/commands/wp';
 import { readAppdata, lockAppdata, unlockAppdata, saveAppdata } from 'cli/lib/appdata';
 import { loadTranslations } from 'cli/lib/i18n';
@@ -98,7 +97,6 @@ async function main() {
 			registerSiteListCommand( sitesYargs );
 			registerSiteStartCommand( sitesYargs );
 			registerSiteStopCommand( sitesYargs );
-			registerSiteStopAllCommand( sitesYargs );
 			registerSiteDeleteCommand( sitesYargs );
 			registerSiteSetHttpsCommand( sitesYargs );
 			registerSiteSetDomainCommand( sitesYargs );
